MultiNetworkPolicy in a separate nav item
