+++ frontend/packages/console-app/locales/en/console-app.json
@@ -62,6 +62,7 @@
   "Routes": "Routes",
   "Ingresses": "Ingresses",
   "NetworkPolicies": "NetworkPolicies",
+  "MultiNetworkPolicies": "MultiNetworkPolicies",
   "PersistentVolumes": "PersistentVolumes",
   "PersistentVolumeClaims": "PersistentVolumeClaims",
   "StorageClasses": "StorageClasses",
@@ -324,10 +325,6 @@
   "Add egress rules to be applied to your selected pods. Traffic is allowed to pods if it matches at least one rule.": "Add egress rules to be applied to your selected pods. Traffic is allowed to pods if it matches at least one rule.",
   "Add egress rule": "Add egress rule",
   "Cancel": "Cancel",
-  "Enable {{kind}}": "Enable {{kind}}",
-  "{{kind}} disabled": "{{kind}} disabled",
-  "Cluster administrator permissions are required to enable this feature.": "Cluster administrator permissions are required to enable this feature.",
-  "MultiNetworkPolicies": "MultiNetworkPolicies",
   "{{path}} is missing.": "{{path}} is missing.",
   "{{path}} should be an Array.": "{{path}} should be an Array.",
   "{{path}} should not be empty.": "{{path}} should not be empty.",
