--- frontend/packages/console-app/src/components/network-policies/network-policy-list/network-policy-list-page.scss
@@ -1,13 +1,0 @@
-.network-policy-list-page.pf-v5-c-tabs {
-  padding-left: 1rem;
-
-  li {
-    padding-top: 2rem;
-
-    button {
-      font-size: var(--pf-v5-global--FontSize--lg);
-      padding-left: 2rem;
-      padding-right: 2rem;
-    }
-  }
-}
