+++ frontend/packages/console-app/src/components/network-policies/useIsMultiNetworkPolicy.tsx
@@ -5,7 +5,10 @@ import { MultiNetworkPolicyModel } from '@console/internal/models';
 const useIsMultiNetworkPolicy = () => {
   const location = useLocation();
 
-  return !isEmpty(location.pathname.match(MultiNetworkPolicyModel.kind));
+  return (
+    !isEmpty(location.pathname.match(MultiNetworkPolicyModel.kind)) ||
+    !isEmpty(location.pathname.match(MultiNetworkPolicyModel.plural))
+  );
 };
 
 export default useIsMultiNetworkPolicy;
